XBE Explorer : Added a list of known address-ranges. Gaps are not detected yet, but manual analysis shows gaps are zero-filled most of the times.

+++ src/DxbxKrnl/uDxbxKrnlUtils.pas
@@ -163,7 +163,7 @@ end;
 procedure SetByteBits(var Bits: Byte; const aIndex: Integer; const aValue: Byte);
 var
   Offset: Byte;
-  Mask: Integer;
+  Mask: Byte;
 begin
   Mask := ((1 shl {NrBits=}Byte(aIndex)) - 1);
   Assert(aValue <= Mask);
